Embed alfa-act/Event into Rule.Event
