+++ packages/alfa-act/src/index.ts
@@ -8,7 +8,6 @@
 export * from "./audit";
 export * from "./cache";
 export * from "./diagnostic";
-export * from "./event";
 export * from "./interview";
 export * from "./oracle";
 export * from "./outcome";
