+++ packages/alfa-act/tsconfig.json
@@ -5,7 +5,6 @@
     "src/audit.ts",
     "src/cache.ts",
     "src/diagnostic.ts",
-    "src/event.ts",
     "src/index.ts",
     "src/interview.ts",
     "src/oracle.ts",
